progress report bugfix, AnalysisSpec pivot signature bugfix, ExecutionValue toString, refresh output on analysis change, refresh output after stop running, offline output info works, run reset action, added LogicExecution.close

+++ kzen-auto-common/src/commonMain/kotlin/tech/kzen/auto/common/objects/document/report/progress/ReportFileProgress.kt
@@ -86,7 +86,7 @@ data class ReportFileProgress(
 
             else -> {
                 val percent = ((readBytes.toDouble() / totalSize.coerceAtLeast(1)) * 100)
-                val percentFormat = percent.toInt().toString() + "." + percent * 10 % 10
+                val percentFormat = percent.toInt().toString() + "." + (percent * 10 % 10).toInt()
 
                 val recentRecordsSpeed = FormatUtils.decimalSeparator(
                     1000L * recentRecordsPerSecond / adjustedDurationMillis)
